Added exit status ios simulator testing
This now uses xcrun simctl to start the app in "wait" mode, then we
attach to the process via lldb and tell it to run. The exit status is
then printed in the lldb stdout. We then parse the stdout and get the
exit code and bubble it up to dinghy.

This may now require root to attach to the process.