Use package name instead of runnable id for dir on target
This means means multiple runnables from the same package (ie
integration tests) will live in the same directory on the target device
and this directory will be stable and will not contain a build hash.

This is way better for disk usage especially when lots for tests
resources need to be copied (as those will not be duplicated on the
target device). This should also improve the first execution time on ssh
devices when there are multiple integration tests in the package as the
resources will need to only be copied once then rsync will do some magic